+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.3.0
 Features:
+* Add [ParametersLogBuilder](http://maven-plugin-utils.projects.gabrys.biz/LATEST/apidocs/index.html?biz/gabrys/maven/plugin/util/parameter/ParametersLogBuilder.html) which allows to log Mojos parameters values
 * Add protected method [AntFileScanner#createDirectoryScanner()](http://maven-plugin-utils.projects.gabrys.biz/LATEST/apidocs/index.html?biz/gabrys/maven/plugin/util/io/AntFileScanner.html)
 * Add protected method [RegexFileScanner#createFileFilter(File, String[], String[])](http://maven-plugin-utils.projects.gabrys.biz/LATEST/apidocs/index.html?biz/gabrys/maven/plugin/util/io/RegexFileScanner.html)